Don't merge in plugin post details menu when it doesn't exist.
Prevents warnings from being occassionally being generated when merging post
detail menu array details together.
The test turns off debug and turns on E_WARNING to prevent the message from
being hidden.  (Which in actual running, it is not on occasion.)
Test also verifies that plugins are added to the menu.

references issue #1756